More work on parallax mapping.
Former-commit-id: e0bdf560f78d91f74e5e642811f13f77e4e939a9

+++ Nu/Nu/OpenGL/OpenGL.PhysicallyBased.fs
@@ -1141,7 +1141,7 @@ module PhysicallyBased =
         // update heights buffer
         let heightsFieldsPtr = GCHandle.Alloc (heightsFields, GCHandleType.Pinned)
         try Gl.BindBuffer (BufferTarget.ArrayBuffer, geometry.HeightBuffer)
-            Gl.BufferData (BufferTarget.ArrayBuffer, uint (surfacesCount * sizeof<int>), heightsFieldsPtr.AddrOfPinnedObject (), BufferUsage.StreamDraw)
+            Gl.BufferData (BufferTarget.ArrayBuffer, uint (surfacesCount * sizeof<single>), heightsFieldsPtr.AddrOfPinnedObject (), BufferUsage.StreamDraw)
         finally heightsFieldsPtr.Free ()
         Gl.BindBuffer (BufferTarget.ArrayBuffer, 0u)
         Hl.Assert ()
